Move KeyValueType to test folder
