+++ web3/src/test/java/org/hiero/mirror/web3/service/ContractCallNestedCallsTest.java
@@ -4,12 +4,14 @@ package org.hiero.mirror.web3.service;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.hiero.mirror.web3.evm.utils.EvmTokenUtils.toAddress;
+import static org.hiero.mirror.web3.service.utils.KeyValueType.CONTRACT_ID;
+import static org.hiero.mirror.web3.service.utils.KeyValueType.ED25519;
+import static org.hiero.mirror.web3.service.utils.KeyValueType.INHERIT_ACCOUNT_KEY;
 import static org.hiero.mirror.web3.utils.ContractCallTestUtil.CREATE_TOKEN_VALUE;
 import static org.hiero.mirror.web3.utils.ContractCallTestUtil.NEW_ECDSA_KEY;
 import static org.hiero.mirror.web3.utils.ContractCallTestUtil.NEW_ED25519_KEY;
 
 import com.google.protobuf.ByteString;
-import com.hedera.services.store.contracts.precompile.codec.KeyValueWrapper.KeyValueType;
 import java.math.BigInteger;
 import java.time.Instant;
 import java.util.ArrayList;
@@ -21,6 +23,7 @@ import org.hiero.mirror.common.domain.entity.EntityId;
 import org.hiero.mirror.common.domain.entity.EntityType;
 import org.hiero.mirror.common.domain.token.Token;
 import org.hiero.mirror.common.domain.token.TokenTypeEnum;
+import org.hiero.mirror.web3.service.utils.KeyValueType;
 import org.hiero.mirror.web3.web3j.generated.NestedCalls;
 import org.hiero.mirror.web3.web3j.generated.NestedCalls.HederaToken;
 import org.hiero.mirror.web3.web3j.generated.NestedCalls.KeyValue;
