+++ web3/src/test/java/org/hiero/mirror/web3/service/utils/KeyValueType.java
@@ -1,0 +1,12 @@
+// SPDX-License-Identifier: Apache-2.0
+
+package org.hiero.mirror.web3.service.utils;
+
+public enum KeyValueType {
+    INVALID_KEY,
+    INHERIT_ACCOUNT_KEY,
+    CONTRACT_ID,
+    DELEGATABLE_CONTRACT_ID,
+    ED25519,
+    ECDSA_SECPK256K1
+}
